First pass isolation of script preprocessor to own pakage. Tests have not yet been moved, to avoid reconfiguring them for running in vscode test
