+++ packages/sl-script-preprocessor/README.md
@@ -1,0 +1,5 @@
+# Secondlife Script Preprocessor
+
+Preprocessing for both Lua and LSL scripts in secondlife.
+
+For a complete guide [see here](doc/preprocessor-guide.md)
