+++ packages/sl-script-preprocessor/src/index.ts
@@ -1,0 +1,14 @@
+/**
+ * @file index.ts
+ * Public index for the sl script preprocessor.
+ */
+export * from "./lexingpreprocessor";
+export * from "./parser";
+export * from "./lexer";
+export * from "./interfaces";
+export * from "./diagnostics";
+export * from "./macroprocessor";
+export * from "./conditionalprocessor";
+export * from "./includeprocessor";
+export * from "./linemapper";
+export * from "./token";
