+++ src/interfaces/configinterface.ts
@@ -6,7 +6,7 @@
  * any direct dependency on VS Code types. All paths are returned as StringUri.
  */
 
-import { StringUri } from './hostinterface';
+import { StringUri } from '#sl-script-preprocessor';
 
 /** Keys used by configuration (mirrors LLConfigNames). */
 export enum ConfigKey {
