+++ src/synchservice.ts
@@ -6,7 +6,7 @@ import * as vscode from "vscode";
 import * as path from "path";
 import * as fs from "fs";
 import { SCRIPT_FILE_PATTERN, ConfigService, NOTECARD_FILE_PATTERN } from "./configservice";
-import { ConfigKey } from "./interfaces/configinterface";
+import { ConfigKey, FullConfigInterface } from "./interfaces/configinterface";
 import {
     ViewerEditWSClient,
     CompilationResult,
@@ -47,10 +47,9 @@ import {
     CommandExecuteResponse,
     CommandListResponse,
 } from "#sl-ide-ws-client";
-import { ScriptLanguage, LanguageService } from "./shared/languageservice";
-import { ScriptIdentity, ScriptSync } from "./scriptsync";
-import { getLanguageConfig } from "./shared/lexer";
-import { HostInterface } from "./interfaces/hostinterface";
+import {LanguageService } from "./shared/languageservice";
+import { buildPreprocessorConfig, ScriptIdentity, ScriptSync } from "./scriptsync";
+import { getLanguageConfig, HostInterface, ScriptLanguage } from "#sl-script-preprocessor";
 import { SyncedFileDecorator } from "./vscode/SyncedFileDecorator";
 import { ObjectContentChangeEvent, ObjectContentService, ObjectTreeChangeEvent } from "#sl-ide-ws-client";
 import { ObjectPinStore } from "./vscode/objectpinstore";
@@ -86,6 +85,7 @@ export class SynchService implements vscode.Disposable {
     private lastActiveChange: number = 0;
     private activeSync: ScriptSync | undefined;
     private host: HostInterface;
+    private config: FullConfigInterface;
     private readonly commandRegistry = new CommandRegistry();
     private initialGenerationDone: boolean = false;
     private pendingLaunchObjectId?: string;
@@ -110,9 +110,10 @@ export class SynchService implements vscode.Disposable {
 
     private disposables: vscode.Disposable[] = [];
 
-    private constructor(context: vscode.ExtensionContext) {
+    private constructor(context: vscode.ExtensionContext, config: FullConfigInterface) {
         this.context = context;
         this.host = new VSCodeHost();
+        this.config = config;
         this.syncedFileDecorator = new SyncedFileDecorator(this);
         // Note: _onDidChangeConnectionState is NOT added to disposables
         // because it must survive activate/deactivate cycles
@@ -125,7 +126,7 @@ export class SynchService implements vscode.Disposable {
                     "SynchService not initialized. Context is required for first initialization.",
                 );
             }
-            SynchService.instance = new SynchService(context);
+            SynchService.instance = new SynchService(context, ConfigService.getInstance());
         }
         return SynchService.instance;
     }
@@ -315,7 +316,7 @@ export class SynchService implements vscode.Disposable {
             syncs.push(...this.findSyncsByTempFilePath(viewerFilePath));
         }
 
-        if(!this.host.config.getConfig(ConfigKey.KeepViewerFileOpen, true) && masterFound) {
+        if(!this.config.getConfig(ConfigKey.KeepViewerFileOpen, true) && masterFound) {
             void closeTextDocument(viewerDocument).catch((error) => {
                 logInfo(
                     `Failed to auto-close viewer document ${viewerDocument.uri.fsPath}: ${error instanceof Error ? error.message : String(error)}`,
@@ -520,7 +521,7 @@ export class SynchService implements vscode.Disposable {
         showStatusMessage("Connecting to Second Life viewer...", handshake);
 
         const port = portOverride
-            ?? this.host.config.getConfig<number>(ConfigKey.NetworkWebsocketPort, 9020);
+            ?? this.config.getConfig<number>(ConfigKey.NetworkWebsocketPort, 9020);
         this.websocket = new ViewerEditWSClient({
             url: `ws://localhost:${port}`,
             logger: {
@@ -536,7 +537,7 @@ export class SynchService implements vscode.Disposable {
                 }
             },
             disconnectDelayMs: (): number =>
-                this.host.config.getConfig<number>(ConfigKey.NetworkDisconnectDelayMs, 1000),
+                this.config.getConfig<number>(ConfigKey.NetworkDisconnectDelayMs, 1000),
         });
         this.context.subscriptions.push(this.websocket.setup(handlers));
         let connected = await this.websocket.connect();
@@ -1175,7 +1176,7 @@ export class SynchService implements vscode.Disposable {
     ) : Promise<vscode.Uri | null> {
         const config =  ConfigService.getInstance()
 
-        const cmt = getLanguageConfig(script.language,config).lineCommentPrefix;
+        const cmt = getLanguageConfig(script.language,buildPreprocessorConfig(script.language, config)).lineCommentPrefix;
 
         if(cmt.length < 1) return null;
