+++ src/test/suite/parse-line-mappings.test.ts
@@ -4,8 +4,8 @@
  */
 
 import * as assert from 'assert';
-import { LineMapper, LineMapping } from '../../shared/linemapper';
-import { filePathToStringUri, StringUri } from '../../interfaces/hostinterface';
+import { LineMapper, LineMapping } from '#sl-script-preprocessor';
+import { filePathToStringUri, StringUri } from '#sl-script-preprocessor';
 import { createMockHost } from './helpers/mockHost';
 import { expectMapping, expectMappings } from './helpers/expectMapping';
